+++ chat-widget/src/components/livechatwidget/common/defaultProps/defaultPersistentChatHistoryProps.ts
@@ -1,7 +1,7 @@
 import { IPersistentChatHistoryProps } from "../../interfaces/IPersistentChatHistoryProps";
 
 export const defaultPersistentChatHistoryProps: IPersistentChatHistoryProps = {
-    pageSize: 4,
+    pageSize: 10,
     dividerActivityStyle: {
         border: "1px solid rgb(96, 94, 92, 0.5)",
         margin: "10px 20%"