+++ chat-widget/src/components/webchatcontainerstateful/hooks/usePersistentChatHistory.ts
@@ -5,9 +5,8 @@ import ChatWidgetEvents from "../../livechatwidget/common/ChatWidgetEvents";
 import { FacadeChatSDK } from "../../../common/facades/FacadeChatSDK";
 import { IPersistentChatHistoryProps } from "../../livechatwidget/interfaces/IPersistentChatHistoryProps";
 import PersistentConversationHandler from "../../livechatwidget/common/PersistentConversationHandler";
-import { TelemetryHelper } from "../../../common/telemetry/TelemetryHelper";
-import dispatchCustomEvent from "../../../common/utils/dispatchCustomEvent";
 import SecureEventBus from "../../../common/utils/SecureEventBus";
+import { TelemetryHelper } from "../../../common/telemetry/TelemetryHelper";
 
 const usePersistentChatHistory = (facadeChatSDK: FacadeChatSDK | undefined, props : IPersistentChatHistoryProps) => {
     const handlerRef = useRef<PersistentConversationHandler | null>(null);
@@ -32,8 +31,6 @@ const usePersistentChatHistory = (facadeChatSDK: FacadeChatSDK | undefined, prop
             
             try {
                 await handlerRef.current?.pullHistory();
-
-                dispatchCustomEvent(ChatWidgetEvents.HIDE_LOADING_BANNER);
                 
                 TelemetryHelper.logActionEventToAllTelemetry(LogLevel.INFO, {
                     Event: TelemetryEvent.LCWPersistentChatHistoryFetchCompleted,